Nettoyage de printemps du package allsky (homogéneisation de Skygen et mode GUI)

+++ src/cds/allsky/BuilderGzip.java
@@ -33,90 +33,83 @@ import cds.tools.pixtools.Util;
 
 
 /**
- * Classe de compression (resp. décompression) de toutes les tuiles Fits
+ * Permet la compression de certaines tuiles Fits (les 3 premiers niveaux et le Allsky.fits)
+ * Rq : Ne compresse jamais le niveau le plus profond
  * @author P. Fernique [CDS]
  * @version 1.0 - mai 2012 - création
  */
-final public class BuilderGzip {
+public class BuilderGzip extends Builder {
    
-   private String root;     // Répertoire racine à partir duquel il faut (un)zipper les tuiles FITS
-   private int verbose;     // Niveau de verbosité 0-rien, 1-étoiles, 2-fichiers
    private int nbFile;      // Nombre de fichires traités
-   private Context context=null;
    
    public BuilderGzip(Context context) {
-      this(context.getOutputPath(),context.getVerbose());
-      this.context=context;
-   }
-   public BuilderGzip(String root,int verbose) {
-      this.root = root;
-      this.verbose=verbose;
+      super(context);
       nbFile=0;
    }
+   
+   public Action getAction() { return Action.GZIP; }
+
+   public void validateContext() throws Exception { validateOutput(); }
    
    /** Gzippe toutes les tuiles FITS ainsi que le fichier Allsky.fits qui se trouve
     * dans le répertoire Allsky repéré par root.
     * Attention: ne change pas pour autant les extensions des fichiers (toujours.fits)
     */
-   public void gzip() { gzipRec(true); }
-
-   /** Gunzippe toutes les tuiles FITS ainsi que le fichier Allsky.fits qui se trouve
-    * dans le répertoire Allsky repéré par root
-    * Attention: ne change pas pour autant les extensions des fichiers (toujours.fits)
-    */
-   public void gunzip() { gzipRec(false); }
-   
-   /** Permutation du mode de compression des tuiles.
-    * Se base sur l'état du fichier Allsky.fits pour déterminer l'état courant */
-   public void switchGzip() throws Exception {
-      MyInputStream mis = new MyInputStream( new FileInputStream( getAllskyPath() ) );
-      boolean doCompress = !mis.isGZ();
-      mis.close();
-      if( verbose>0 ) System.out.println("Starting "+(doCompress?"gzip":"gunzip")+"...");
-      gzipRec(doCompress);
+   public void run() throws Exception { 
+      gzipRec(true);
    }
    
-   // retourne le path du fichier Allsky.fits
-   private String getAllskyPath() { return root+FS+"Norder3"+FS+"Allsky.fits"; }
+   public boolean isAlreadyDone() {
+      if( !context.actionPrecedeAction(Action.INDEX, Action.TILES)) return false;
+      if( !context.actionPrecedeAction(Action.TILES, Action.GZIP)) return false;
+      if( context.actionAlreadyDone(Action.GUNZIP) && !context.actionPrecedeAction(Action.GZIP, Action.GUNZIP)) return false;
+      context.info("GZIP seems to be already done");
+      return true;
+   }
    
    // lance le gzip (resp gunzip) récursivement sur tous les répertoire Norder??
    // Dans le cas où un fichier est déjà gzippé (resp. gunzippé), le fichier est simplement ignoré
-   private void gzipRec(boolean compress) {
+   protected void gzipRec(boolean compress) throws Exception {
+      String path = context.getOutputPath();
+      int maxOrder = Util.getMaxOrderByPath(path);
       int order;
-      String path = root;
       
       // Parcours de tous les répertoire Norder?? trouvés
       for( File nOrder : (new File(path)).listFiles() ) {
+         if( context.isTaskAborting() ) throw new Exception("Task abort !");
          String name = nOrder.getName();
          if( !name.startsWith("Norder") ) continue;
          if( !nOrder.isDirectory() ) continue;
          try { order = Integer.parseInt(name.substring(6)); }
          catch( Exception e ) { continue; }
          
-         // On ne compresse pas les tuiles au-delà de l'ordre 5
-         if( order>5 && compress ) continue;
-         
          // traitement particulier pour le fichier Allsky.fits qui se trouve dans le Norder3
          if( order==3 ) {
-            String allsky = getAllskyPath();
-            if( (new File(allsky)).isFile() ) gzip(allsky,compress);
+            String allsky = path+FS+"Norder3"+FS+"Allsky.fits";
+            if( (new File(allsky)).isFile() ) { gzip(allsky,compress); nbFile++; }
          }
+         
+         // On ne compresse pas les tuiles au-delà de l'ordre 5
+         // Ni celles du dernier niveau
+         if( compress && (order>Constante.GZIPMAXORDER || order==maxOrder) ) continue;
          
          // Traitement de toutes les tuiles du niveau
          long maxNpix = Healpix.pow2(order);
          maxNpix = 12*maxNpix*maxNpix;
          for( long npix=0; npix<maxNpix; npix++ ) {
             String filename = Util.getFilePath(path, order, npix)+".fits";
-            if( (new File(filename)).isFile() ) gzip(filename,compress);
+            if( (new File(filename)).isFile() ) {
+               gzip(filename,compress);
+            }
          }
       }
-      if( verbose>0 ) System.out.println();
-      if( nbFile==0 ) System.out.println("No tile found !");
+      if( nbFile==0 ) context.warning("No tile found !");
+      
    }
    
    // gzip (resp. gunzip) du fichier indiqué. 
    // Dans le cas où un fichier est déjà gzippé (resp. gunzippé), le fichier est simplement ignoré
-   private void gzip(String file,boolean compress) {
+   private void gzip(String file,boolean compress) throws Exception {
       try {
         File in = new File(file);
         if( !in.isFile() ) throw new Exception(file+" does not exist !");
@@ -135,33 +128,27 @@ final public class BuilderGzip {
         
         byte [] buf = new byte[8192];
         int n;
-        while( (n=mis.read(buf))>=0 ) mos.write(buf,0,n);
+        while( (n=mis.read(buf))>=0 ) {
+           mos.write(buf,0,n);
+           if( context.isTaskAborting() ) break;
+        }
         
         mos.close();
         mis.close();
         
-        in.delete();
-        out.renameTo(in);
+        if( context.isTaskAborting() ) out.delete();
+        else {
+
+           in.delete();
+           out.renameTo(in);
+
+           nbFile++;
+           if( context!=null ) context.setProgress(nbFile);
+        }
         
-        if( verbose>1 ) System.out.println((compress?"gzip ":"gunzip ")+file);
-        else if( verbose>0 && nbFile%100==0 ) System.out.print("*");
-        
-        nbFile++;
-        
-     } catch( Exception e ) {
-        if( verbose>0 ) System.err.println(e.getMessage());
-     }
+     } catch( Exception e ) { }
+     
+     if( context.isTaskAborting() ) throw new Exception("Task abort !");
    }
    
-   // Test
-   static public void main(String [] argv) {
-      try {
-         long t = System.currentTimeMillis();
-         BuilderGzip bdMoc = new BuilderGzip("C:/HalphaNorthALLSKY",1);
-         bdMoc.switchGzip();
-         System.out.println("Done in "+((System.currentTimeMillis()-t)/1000)+"s !");
-      } catch( Exception e ) {
-         e.printStackTrace();
-      }
-   }
 }
